Extraer UI de estado de sincronización de mark.js
Tercer paso de la descomposición de mark.js: setOfflineBanner(),
updateSyncStatus() y refreshSyncStatus() se mueven a
mark/sync-status-ui.js. setOfflineBanner() ahora recibe checkEnableMark
como callback en vez de leerlo del closure de mark.js (mismo efecto,
llamado en los mismos 3 puntos). Refactor puro, sin cambios de
comportamiento.
